Updated Test.clp to contain more files to load
I do need to make the caveat that it takes a while to load these files. I think
it's because of a bug in CLIPS. I'm not sure. Recompiling the program with -O2
for each file cut the running time down greatly but it's still taking a very
long time for really no reason. It's the act of sending a message that causes
the slowdown...When I do (send ?target put-FileContents ?contents) CLIPS will
lock itself for a little bit and I'm not sure why :(