Added a machine description class
The objective of this class is to allow the programmer to define the basic
layout of a given machine to be used for cases where a binary file format is
actually the values of a given struct written to a file.

In these cases the most ideal way to reverse it is by figuring out potential
layouts. Since there are so many different kinds of layouts it is up to the
programmer to assist the expert system by doing a little bit of manual analysis
and identifying any interesting points for it.

If the reverser can provide any extra information it will be helpful to the
expert system in assisting it.

For instance, if the programmer can give an idea of what the structure consists
of then it will make life easier.