Put the File class into its own file.
I added several rules and message handlers.

I also added a file type for the DOS MZ format as the first supported format.
It requires that the programmer already figured out that the file is of that
format. I have not figured out a safe way to do file type identification.